chore: infomodel test
