+++ .flake8
@@ -4,6 +4,6 @@ docstring-convention = google
 max-complexity = 10
 ignore = ANN101,B009,E203,E501,W503,DAR101,DAR201,D100,D103,D107,D101,D104
 max-line-length = 120
-per-file-ignores = __init__.py: F401,tests/*:S101
+per-file-ignores = __init__.py: F401,tests/*:S101,B950
 application-import-names = fdk_harvester_bff, tests
 import-order-style = google
