+++ docker-compose.yml
@@ -5,6 +5,7 @@ services:
     image: eu.gcr.io/digdir-fdk-infra/fdk-harvester-bff:test
     environment:
       - DATASET_HARVESTER_BASE_URL=http://fdk-services-mock:8080
+      - INFORMATION_MODEL_HARVESTER_BASE_URL=http://fdk-services-mock:8080
       - ORGANIZATION_CATALOGUE_BASE_URI=http://fdk-services-mock:8080
       - REFERENCE_DATA_BASE_URI=http://fdk-services-mock:8080
     ports:
